Improve included run scripts and properties
This uses the gosh.args properties to avoid running an interactive
shell. The reason for this is that when running under process
supervision, the lack of data on stdin causes the local shell to
terminate and the framework to be shut down as a result.

Set the default log level to 1.